用户登录

+++ app/models/conversation.go
@@ -1,9 +1,9 @@
 package models
 
 type Conversation struct {
-	Id      int64  `json:"id"`       //编号
-	UserAID int64  `json:"userA_id"` //用户A的ID
-	UserBID int64  `json:"userB_id"` //用户B的ID
+	Id      int  `json:"id"`       //编号
+	UserAID int  `json:"userA_id"` //用户A的ID
+	UserBID int  `json:"userB_id"` //用户B的ID
 	Content string `json:"content"`  //内容
 	Time    string `json:"time"`     //时间
 }
